03. DAL/NewHabr.DAL/Repository/NotificationRepository.cs
using System;
using Microsoft.EntityFrameworkCore;
using NewHabr.DAL.EF;
using NewHabr.Domain.Contracts.Repositories;
using NewHabr.Domain.Models;

namespace NewHabr.DAL.Repository;

public class NotificationRepository : RepositoryBase<Notification, Guid>, INotificationRepository
{
    public NotificationRepository(ApplicationContext context) : base(context)
    {
    }


    public async Task<Notification?> GetByIdAsync(Guid notificationId, Guid userId, bool trackChanges, CancellationToken cancellationToken)
    {
        return await GetById(notificationId, trackChanges)
            .Include(n => n.Users.Where(u => u.Id == userId))
            .FirstOrDefaultAsync(n => n.Users.Any(user => user.Id.Equals(userId)), cancellationToken);
    }

    public async Task<ICollection<Notification>> GetUserNotificationsAsync(Guid userId, bool unreadOnly, bool trackChanges, CancellationToken cancellationToken)
    {
        var query = GetAll(trackChanges);

        if (unreadOnly)
            query = query.Where(n => n.IsRead == false);

        var notifications = await query
            .Include(n => n.Users.Where(u => u.Id == userId))
            .Where(n => n.Users.Any(user => user.Id.Equals(userId)))
            .ToListAsync(cancellationToken);

        return notifications;
    }
}

